Add files via upload
In this project, I've used intermediate to advanced level SQL queries to explore the employee database of a company.  For demonstration purposes, I created a employee database with multiple tables and around 10-15 records per table.  I've used various SQL techniques such as Joins, Sub-queries, CTEs, Text functions and so on to analyse the data.